+++ packages/extension/src/libs/nft-handlers/types/unisat.ts
@@ -1,0 +1,14 @@
+export interface UnisatNFTType {
+  inscriptionId: string;
+  output: string;
+  preview: string;
+  inscriptionNumber: string;
+  location: string;
+}
+
+export interface UnisatResponse {
+  code: number;
+  data: {
+    list: UnisatNFTType[];
+  };
+}
